patch: do not overwrite highest version in dependencies settings
This is a hotfix that resolve an issue where dependencies overwrote the
version number of packages that also got updated. e.g.

I apply a ‘minor’ (0.0.0 → 0.1.0) version on the Foo package which
depends on Bar. Bar has no update applied on it, it’s version
remains the same (0.0.0), yet because of the depencies the Foo package
version is set to match the Bar package (0.1.0 → 0.0.0).

Optimally rather than validating each version at the last step (right
before the write) we would compare the version early and only send them on
the “write” pile if they are correct.